Try reading fmap or fdt from file first
With larger flash chip sizes and slow interfaces, looking up the
onboard flash for fmap or fdt could take very long time, especially if
the onboard flash is not programmed.

On the other hand, the file containing the image to be programmed
pretty much always contains fmap or fdt.

This patch modifies the interface of the search subsystem to pass it a
callback function to search the image instead of passing it the flash
structure pointer.

The search function first is called with the callback to search in the
supplied image file, and then if the file was not supplied or neither
map nor fdt were not found - try again, this time looking in the chip
itself, using a callback which reads in the flash.

BRANCH=none
BUG=b:79492818
TEST=verified that invoking flashrom with -w and -i, when trying to
     program a fully erased flash chip, does not hang for a very long
     time while fmap is searched, as it is retrieved from the supplied
     file.

Change-Id: I50bfe729c2290c06136a1cc3a1ca91f59d42cc91
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1053405
Reviewed-by: Simon Glass <sjg@chromium.org>