add option to disable diffing
In some cases the programmer clocks the flash chips at low clock rate,
reading the chip contents usually done to be able to diff it with the
new contents, takes very long time.

It would be beneficial just erase the chip first and then program it
without reading. This patch adds a new long command line option
'--do-not-diff' to allow user to enforce this mode.

Also, in case user requests '--fast-verify' - let's verify only the
parts written during this flashrom invocation instead of verifying the
entire chip even when a single section was in fact written.

BRANCH=none
BUG=b:79492818
TEST=took a Scarlet device through the following steps:
   - erase the chip using -E command line option
   - program it with --do-not-diff option, observe it boot into chrome os

Change-Id: I9bdc41b5f0e62edda8beb0f73a7f761e84628578
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1068269
Reviewed-by: Simon Glass <sjg@chromium.org>