Adjust start of action descriptor comparison
When looking for blocks with differences, the first character of each
block was missed because the value was auto-incremented by the loop. If
the first character was the only character in a block that changed, the
block wouldn't be flashed, causing a miscomparison during the verify
routine.

BUG=b:119158699
TEST=Flash grunt image where only the first character of the block
is different.
BRANCH=None

Change-Id: Ifa85e800f727ea2709512a8779532758599f3c31
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1339659
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Kevin Chiu <kevin.chiu@quanta.corp-partner.google.com>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Kevin Chiu <kevin.chiu@quanta.corp-partner.google.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>