Add support for the ENE Embedded Debug Interface EDI and KB9012 EC
This is a backport of `commit 80ae14e5` upstream;

The ENE Embedded Debug Interface (EDI) is a SPI-based interface for
accessing the memory of ENE embedded controllers.

The ENE KB9012 EC is an embedded controller found on various laptops
such as the Lenovo G505s. It features a 8051 microcontroller and
has 128 KiB of internal storage for program data.

EDI can be accessed on the KB9012 through pins 59-62 (CS-CLK-MOSI-MISO)
when flash direct access is not in use. Some firmwares disable EDI at
runtime
so it might be necessary to ground pin 42 to reset the 8051
microcontroller
before accessing the KB9012 via EDI.

The example of flashing KB9012 at Lenovo G505S laptop could be found
here:
http://dangerousprototypes.com/docs/Flashing_KB9012_with_Bus_Pirate

Signed-off-by: Mike Banon <mikebdp2@gmail.com>
Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Reviewed-on: https://review.coreboot.org/23259
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>

BUG=chromium:478356
BRANCH=none
TEST=still builds

Change-Id: I870eea53dcfce2cc773f4b04594542c329f437b7
Signed-off-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1484921
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>