Align with upstream chip vs flashctx distinctions
In the upstream `commit 5a7cb847f` the 'struct flashchip' became
a field in the 'struct flashctx' instead of a entire copy.

While the internal tree had most of the changes in this commit,
not all of them exists causing needless miss-alignment of the
two source trees.

V.2: Rebase on rework of previous commit.
V.3: Rebase on master.

BUG=chromium:478356
BRANCH=none
TEST=still builds

Change-Id: I734df44817379135f7a965dde9f399d1540d04dd
Signed-off-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1484922
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>