dediprog: Add protocol variant support for newer SF600 fw revs
This is a backport of the upstream `commit f73f8a732`
 dediprog: implement command spec for firmware >= 7.2.30

However this also contains elements of `commit ae61651bfa`
 dediprog: support new communication protocol, cleanup and enable by
 default
that was only partially applied to our tree. Although
the backport here is not a straight forward one were we have
cherry-picked hulks where it makes sense to do so to get a
viable itermediate patch working.

BUG=chromium:478356
BRANCH=none
TEST=still builds

Change-Id: Idf960bc7c57129e9f63e4aaeef6c9b82b0339555
Signed-off-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1492816
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>