Fix check for EC region read/write access permissions
When new SPI regions were added there was not a corresponding register
added to the SPI MMIO region that allowed for an easy way to read the
access permissions of the new regions.

The original register is only 1 byte and can only handle the first 7
(0 based) regions, while the EC is region 8.

Instead use the Flash Descriptor Observability interface to read out
the FLMSTR registers directly from the descriptor and check the access
permission bits in there for the EC region.

This only works on controllers that have the new FLMSTR layout.

BUG=b:130706387
BRANCH=none
TEST=ensure that flashrom can tell when EC region is read-only

Change-Id: I3b575494ad4cb72be3449ddb20f4f891e380ca77
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1584649
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>