ichspi: Set EC region as RW if descriptor override is enabled
If the flash descriptor override pin is asserted then the flash descriptor
security is disabled.  For most regions this is represented in the flash
region status register, but since the EC does not have a defined bit in
that register it looks at the descriptor directly and so it sees that
the region is marked as read-only and treats it as such.

In order to support flashing this region if descriptor override is enabled
the code should first check for that state and mark the EC region as
read-write if that is the case.

BUG=b:133642633
BRANCH=M75
TEST=assert flash descriptor override and flash EC region with flashrom

Change-Id: I2d1986c87a0f9284ae5f2fb8f4aab2a4961ee9ef
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1632513
Tested-by: Chris Zhou <chris_zhou@compal.corp-partner.google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>