BACKPORT: dediprog.c: Add id parameter to dediprog programmer
When multiple dediprog programmers are connected, the 'id' parameter
allows you to specify which one to use. The id is a string like SF012345
or DP012345. The value is printed on a sticker on the back of the dediprog.

This is an improvement over the 'device' parameter which is based on
enumeration order and changes when you plug/unplug devices or reboot the
machine.

To find the id without the sticker, run flashrom with the -V option.
This prints the ids as they are enumerated.  Alternatively, with dpcmd,
you can use the --list-device-id and --fix-device commands to list and
write device ids respectively.

Note this only supports SF100 at the moment, but SF600 support is
possible with more work.

BUG=none
BRANCH=none
TEST=none

Original-Change-Id: I4281213ab02131feb5d47bf66118a001cec0d219
Original-Signed-off-by: Ryan O'Leary <ryanoleary@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/34160
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
(cherry picked from commit 301ae22b456a040d8944daa0268aa97aa93a517f)

Change-Id: Ibaad68f01463606c908d881b9994b02f62fc4774
Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2556719
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>