flashchips.h: merge chip id MACRONIX_MX25U12835E
This chip is named MACRONIX_MX25U12835E in upstream's flash.h but was
named MX25U12835F (differs in final character of ID) in Chromium's fork
of flashrom. As part of merging with upstream, use the same id as
upstream.

Also adjusts uses of the MACRONIX_MX25U12835F identifier in
- flashchips.c
- writeprotect.h

BRANCH=None
BUG=chromium:478356
TEST=Built flashrom and ran flashrom -L

Change-Id: If28da79be51fc770ef9a212e9e397b12dd811441
Signed-off-by: Alan Green <avg@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1683555
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>