flashchips_parser: Add parser code
Parse and manipulate flashchips.c files, for the purpose of bringing
together the upstream and Chromium versions.

This file is submitted for documentation purposes. As noted in the file
comments, it is not intended to be maintainable nor to have a lifetime
beyond re-joining the fork.

BRANCH=None
BUG=chromium:478356
TEST=Manual testing

Change-Id: I7c10a449b9b3545e08b9ed23c3dc7a745096e77a
Signed-off-by: Alan Green <avg@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1687632
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>