spi: Drop our implementation of spi_unbound_read()
With only minor adjustments to spi_read_chunked() that
aligns us with upstream we do not need a seperate path
and feature flag. Thus, this mechanism is obsolete in
practice after the previous commit.

The implementation of 'spi_read_unbound()' is in a sense
essentially a generalisation of 'spi_read_chunked()'.
The difference is that the bound of 'spi_read_unbound()'
is the granularity 'chunksize' while in the case of
'spi_read_chunked()' the outter loop bounds us to the
min() of either 16MBytes or total_size expressed as
'area_size' and read each area separately.

V.2: squash in,
 - flashchips.c: Drop FEATURE_UNBOUND_READ flag

BUG=chromium:478356
BRANCH=none
TEST=still builds

Change-Id: Ic3549b46ff2a62319d3b2c8c12ad589833fc702d
Signed-off-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1720550
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>