avl_tool: canonicalize parse_crosssystem() with unit test
Also converting a panic if wpsw_cur doesn't appear in the provided
string into a reported error, further saving allocation of a Vec and
checking for trailing whitespace on the tag to ensure we can't be fooled
by a tag that merely has `wpsw_cur` as a prefix.

V.2: rebased on master.

BUG=b:140001966
BRANCH=None
TEST=still builds

Change-Id: Ia361d5bb92225eea67ccdcb2453ec65ce4bd0f1d
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1768392
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>