avl_tool: Introduce a TestEnv and RAII for WP
The TestEnv controls sequencing and attempts to ensure the system
remains in a consistent state after running tests. It also provides a
way to expose state to individual tests.

To provide a way to control the flash write protection, a
WriteProtectState is a RAII handle to the current hardware state. Tests
can manipulate this and avoid unnecessary operations because it compares
the current state with the intended state, and when a WriteProtectState
gets dropped it returns the system to its original state.

The tests are unchanged here, except for the flash sanity test which is
deleted because it is implemented in the runner. A small shim supports
running the old tests.

BUG=b:141142893,b:140891473
TEST=Ran tester on hardware successfully
BRANCH=None

Change-Id: I3bc3738f2b1b98738745b571a8edb9548f86ac31
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1806637
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>