avl_tool: more intelligently find the chip name
flashrom recently began emitting more than just a "coreboot" line and
the requested output, so we'll look for a line with the expected format
and only return the information from that, rather than every other line
smashed together.

This also means we can get rid of utils::vgrep, since this was the last
user of it.

BUG=b:140526168
BRANCH=None
TEST=Unit tests pass

Change-Id: I8856e1f214de9f5d5ccff634b9653b9b467463d2
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1792688
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>