avl_tool: support logging to a file
This isn't supported by env_logger, so we drop that dependency and
instead implement a simple logger ourselves.

To ensure output that should be visible remains visible, some log
statements are converted to prints to stderr and some prints are changed
to logs.

v2: move the logger to its own module, add tests for it

BUG=b:140841859
TEST=Manually exercised --log-file and -d options, inspected logs.
BRANCH=None

Change-Id: I8869543c27410d0ae22a6de67a5005ad86b125d8
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1798230
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>