avl_tool: support specifying tests on the command line
We accept the names of tests as optional positional arguments, executing
only those tests if any names are specified (and all tests otherwise).
If a name is specified that doesn't match any tests, a warning will be
logged. The names are case-insensitive.

BUG=b:140977910
TEST=cargo test
BRANCH=None

Change-Id: Ia27e88ff8a5907566afe144ac0979b15db5cde3e
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/1815447
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>