writeprotect.c: Verify updated value of status register with mask
Some flashchips would retain the WEL bit of status register after some
flash commands when SWWP / HWWP are all enabled and accessed array is in
the protected area.

As a result, once WEL bit is retained and SWWP is disabled, the next
flash command will see original value of status register with WEL bit
but the updated value of status register will be without WEL bit.
Finally this call will be judged as failed because of mis-matching
between expected value and result.

The solution here is to build what bits in the status register should be
checked by looking for the info in status_register_layout of generic_wp.
This is similar to w25q series actually which defined the constant -
MASK_WP_AREA statically; reasons of not just leveraging w25q series are
1. w25q series don't consider the case of 5 BP bits yet so need to
introduce a new struct for this case. Instead of introducing new one
based on w25q, improving generic_wp is more scalable way to support
various BP bits. 2. w25q series assume sec and tb bits in the status
register but we don't need them in this case actually so the generic_wp
is more proper for this case.

BUG=b:146996879
BRANCH=none
TEST=manually enable / disable SWWP when HWWP is enabled / disabled and
check whether the difference of WEL bit affects the result to be good or
not.

Change-Id: I8d93645de76b06706d772ff2d565226d478a3bf8
Signed-off-by: Marco Chen <marcochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2050045
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>