linux_spi.c: remove manual_mknod()
This was added in `282c24135` as a workaround for spi devices not
being created at chromeos_startup on Snow. We are removing it because
it's a brittle solution to a problem that should be solved by fixing
the device configuration instead.

No currently supported devices are known to have this issue. If anyone
still needs to manually create spi devices, they can still manually
create the device with mknod before running flashrom.

BUG=b:153598437
BRANCH=none
TEST=`sudo flashrom -p linux_spi` successfully runs on Hana

Signed-off-by: Nikolai Artemiev <nartemiev@chromium.org>
Change-Id: Ia5a111fc82224d924e2f625868d871ba770463ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2206525
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>