CHROMIUM: read fmap_base from ACPI/FDT directly
Previously, this function shelled out to crossystem for reading the
main firmware FMAP base address.  This utility is only available on
Chrome OS, so it's unlikely upstream flashrom will want this code.

However, if we read directly from ACPI (for x86) or FDT (for ARM), we
can avoid shelling out to crossystem, which will certainly look less
like garbage to upstream.

Note that none of the code this touches exists in upstream, and
there's no equivalent feature there that we can backport yet, so it's
best to land this patch in our own local copy, and we can see as this
code cleans up a little how well we can adapt the functionality to
upstream.

BUG=b:157897361
BRANCH=none
TEST=On octopus (x86) and scarlet (ARM), run:
         "flashrom -V -r -i FMAP:/tmp/fmap.bin"
     - Observe computation of correct fmap_base in debug output
     - No differences to /tmp/fmap.bin

Change-Id: Ie4df4a20fb92bf37fcff395a00fa941fa358d7e1
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2222844
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>