linux_spi.c: remove sysfs walker for finding spi devices
We are removing the sysfs hack since it doesn't work on all boards and makes too many assumptions about device configuration to be useful in upstream. It should be possible to support all boards with only the device tree search.

Some history:

Before this patch, linux_spi would search the device tree for a spi device to use if one wasn't specified. If that failed, linux_spi would fall back to searching sysfs. The sysfs search was added first, in `a20b39347b9750c686b2a633e84e0e93b8e6cbc2`. It attempted to pick the right spi device by finding one that had a corresponding MTD device. Quoting from the commit message:

> We made an assumption here that the SPI flash would also export its MTD interface too. So that we traverse all /dev/spidevX.0 and see if /dev/spidevX.1 is a MTD device.

However, this filtering was not sufficient to find the right device on danger, prompting the addition of the device tree search in `2d06918c3afea8ced513cf0815b519eef53d92f5`. The sysfs search was kept as a fallback, but it's not clear whether that was because some devices didn't work with the device tree search, or just to reduce the chance of breaking something.

BUG=b:153598437
BRANCH=none
TEST=builds

Signed-off-by: Nikolai Artemiev <nartemiev@chromium.org>
Change-Id: Icf2f0c4d9335a0a02e683aa8c79ebad87f9e886f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2241063
Tested-by: Nikolai Artemiev <nartemiev@google.com>
Commit-Queue: Nikolai Artemiev <nartemiev@google.com>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>