flashrom,layout: Dr. Strange chimeric combination of multiple timeslines
This is a highly chimerical commit that has foundations in the upstream
commit,

 * 8268fdb9 - layout: Verify layout entries before building a new image using them

and the downstream commit,

 * 671fa135c - Add bounds checking to handle_romentries()

and some other additions to try and bring layout and flashrom to be
more coherent with the timeline of upstream. This allows more upstream
work to be cherry-picked in and action_descriptor to be more closer to
upstreamable.

BUG=b:139101642,b:162158026
BRANCH=none
TEST=builds

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ibe2257e542bcd4d95cca9d1e397e508ae912a90b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2319506
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>