fmap/layout: Peal back fmap_find() into just a wrapper fn
This inlines the logic of fmap_find() fancy callback logic
to be unrolled into the one call site in layout.c and so
by doing corners out our differences in layout.c to the
one file. fmap_find() isn't entirely inlined as is_valid_fmap()
is static-local to fmap.c and we wish to keep it this way
at the moment.

BUG=b:162047014
BRANCH=none
TEST=builds

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I15c7d778b47ace9d8b1ec6e2bb5847832f69a2d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2361365
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>