linux_spi.c: align argument parsing with upstream and fix memory leaks
Reorder argument parsing and free() strings returned by
parse_programmer_param() to match upstream. To reduce the
complexity assiciated with freeing the device filename string,
check_fdt() and check_sysfs() now return heap allocated strings.

This commit is mostly a repeat of `ac7ed8b70847`, which was
reverted in `5556b94dedcb` because it was built on top of a commit
that needed to be reverted, but did not cause a regression itself.

An additional difference between this and the original patch
is that check_sysfs() has been changed to return a heap-allocated
string in this patch, whereas it had been deleted in the original
patch. Also, malloc/sprintf have been changed to calloc/snprintf
as suggested by quasisec@.

BUG=b:153598437
BRANCH=none
TEST=builds

Signed-off-by: Nikolai Artemiev <nartemiev@chromium.org>
Change-Id: I71f94ce0598d3b3ff24c789d6844e03d6df7ba8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2383220
Tested-by: Nikolai Artemiev <nartemiev@google.com>
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>