programmer.h: Ensure mec1308_init symbol is available
mec1308_init() is the main entry-point and so the prototype is required
for the dispatch table. However, currently internal.c layer violates and
calls mec1308_probe_spi_flash() directly to 'auto-detect' the EC, this
later needs to be removed but leave that symbol in the header for now
for that reason.

BUG=b:170189476
BRANCH=none
TEST=builds

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: If75af3b5ebc24450ac5fb2add19cd1de81cc2713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2458938
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>