Makefile: Sync script func and scripts
Unfortunately this is a large commit in order to be
atomic, such is the nature of gnu-make++bash.

The aim here is to resync our Makefile with upstreams
and the focus of this commit is specifically around
the various scripts and func that burn in version numbers
and check OS features.

Squash in,
  util/getrevision.sh: Take upstreams version

  util/z60_flashrom.rules: Import from upstream
  util/list_yet_unsupported_chips.sh: Import from upstream
  util/flashrom_partial_write_test.sh: Import from upstream
  util/git-hooks: Import from upstream

BUG=b:171014265
BRANCH=none
TEST=emerge and test on DUT that cli gives correct strings.

Disallow-Recycled-Builds: test-failures
Cq-Depend: chromium:2491623
Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ieb34bb7b03cec62ba7846b1ee180686a0cb5c41d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2477704
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>