flashrom.c: Fix 'internal'/'host' assumption of map_flash() failures
Unfortunately some, but not all, board chipset+spi chip combos
result in map_flash() failures while other configurations require
it. Previously the mapping was done unchecked and so a failure in
mapping was silient however now that maps are done via a dedicated
function from upstream and the return value is getting checked it
became apparent some board topologies fail to map_flash for some
reason.

For now we restore previous behaviour by not checking the return
value and instead proceed assuming the map_flash() was successful.
This results in a nasty looking msg in logs however the msg is
correct and useful as a data collection mechanism to find out how
wide spread this is and to dig down what is common, thus we leave
the msg in for now.

BUG=b:171093672,b:170666312,b:171229584
BRANCH=none
TEST=verified to fix Volteer && Trogdor.

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I1d01654c56a65f4b86eee265a608fc5e23aa4f50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2486720
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>