util/flashchips_parser: Kill it off
This has now out-lived it's intended purpose to dedup flashchips.c
Since this code was not intented to be written for long term
maintainability we just delete it and leave it to history.

Moving forwards the intent is for patches to land in upstream
first and flashchips.c to only be a mirror of upstream.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ifce0a9800fde8ad2c821352c3812cfc4f6af971d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2491901
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Alan Green <avg@chromium.org>
Reviewed-by: Alan Green <avg@chromium.org>