programmer.h: spi_master_4ba() fix to return true
Previously the spi master support for 4BA was never
interrogated however now that we make use of this inline
we are validating that the spi driver does indeed support
this feature of the spi flash chip. The issue is that the
re-enterent patch stream has now landed yet and so the
relevent details of the spi master are not within the
flashctx->mst addr space. This makes the control flow of
flashrom.c to mistakenly assume that 4BA would then not
be supported.

Issue first seen after commit 9997445971cd7df721db228deac6dd3309039458

BUG=b:171755805
BRANCH=none
TEST=validated on 'volteer' with,
 - Found chipset "Intel Tigerlake".
 - Found GigaDevice flash chip "GD25Q256D"

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I27ee2e9c51c19097a5259ab5cf73a4be31bb9b0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2500808
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>