Makefile: Sort out building Flashrom as a lib
Here we divide out, as per-upstream, the cli objects
from the Flashrom core functionality objects. This
allows for the core logic of Flashrom to be liberatized
and the cli component to link to it.

The change should be a NOP, amounting more or less to a
reshuffle to match upstreams Makefile with the exception
of a few extra files for some Spansion chips we support
and action_descriptor to both be also included in the core.

BUG=b:171014265
BRANCH=none
TEST=builds

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I5337cff0fa05a4468aafaa0a2ab5525af6c139f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2505312
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>