ichspi.c: Restore previous behaviour of check_fd_permissions_hwseq()
Workaround a regression from the commit
69c3a0df8cc39c876dddfd26a490d7da54533a43.

check_fd_permissions_hwseq() && check_fd_permissions() were meant
to be dedup however there is a slight difference in logic which causes
issue when the descriptors are not correct for the chipset in
chipset_enable.c. Since our chipset_enable.c is slightly different
from upstream and we have abused some chip wrapper callbacks as other
families this causes issues with the descriptors. This winds up breaking
things when the stars align between this commit and the nature of our
chipset_enable.c

This only turns up in very specific cases when the firmware updater is
writing in a very specific pattern to SPI flash. The issue cannot
necessarily be repro'ed simply doing a full erase then write cycle,
atleast not reliably.

BUG=b:171892105,b:156870835,b:155142726,b:139255749,b:139258259
BRANCH=none
TEST=repro on Edgar and resolve's issue.

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I48617c9ebf28cc6cc44d98e52688d7611a453622
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2517210
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>