sb600spi.c: Drop half-done ACPI magic
This was implemented based off a partial data sheet. The rest
is either unknown or under NDA and upstream manages just fine
without it. Just drop it and leave it to history, if it is
required for some purpose a full implementation would be needed
for upstream any way.

BUG=none
BRANCH=none
TEST=tested on Dalboz && Grunt

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ia9a59b1af6f3f1588e5f25ac10953544b0a708ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2554663
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>