spi4ba.c: Delete local bespoke 4BA support
There is no indication this works at all as non-x86 platforms
do not have such large SPI rom's hooked by the 4ba callback
and in the case of ichspi when the chip definition is copied
the callback is not and therefore is NULL within the chip.
Therefore this code is effectively never called.

The following chips are impacted by this change:

 - GD25Q256D
 - MX25L25635F/MX25L25645G
 - W25Q256.V
 - W25Q256JV_M

BUG=b:170690915
BRANCH=none
TEST=builds

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ic4c346cdc64366e6efe45cf862e223ee253f8ff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2562867
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>