programmer.c,ichspi.c: Sync closer to upstream
This allows programmer.c to align with upstream and ichspi.c
to be slightly closer. This is used in action_descriptor.c
so cannot be made static at the moment due to the tech debt.

V.2: action_descriptor needs to be modified as it uses ichspi
 with an unknown chipset to more or less indicate its running
 on x86. This is a very poor design however we try not to address
 the issues of action_descriptor here.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ideb81b32c695de059225e808918fcc4773b974f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2562874
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>