dediprog.c: Fix probe_flash() logic for autovoltage feature
Since the tree is now reentrent we need to be explicit about
which spi master is intended to be used for the probe_flash()
call. Obviously since we are calling the probe_flash() from
within the dediprog spi master we can assume it is the correct
master and therefore we prepare an encoded registered master
to pass into the call to satisfy the signatures requirements.

BUG=b:178554528
BRANCH=none
TEST=none

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: Ib613ef964444cdc1a35df5d679da9d848f42e176
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2651771
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>